+++ tests/vfs.lua
@@ -6,18 +6,27 @@ local N = require("namespace")
 local core = N:GetNamespaceCore()
 local vfs = core.FS:GetProvider("vfs")
 
-
-
+-- Basic write
 vfs:AddDirectory("Test")
-vfs:AddFile("Test._", "return {Name = \"Test\"}")
-vfs:AddFile("Test.Hello", "(...).HelloFlag = true; return {Name = \"Hello\"}")
+vfs:AddFile("Test._", [[ return {Name = "Test"} ]])
+vfs:AddFile("Test.Hello", [[ (...).HelloFlag = true; return {Name = "Hello"} ]])
 
+vfs:AddDirectory("Test.Embedded")
+vfs:AddFile("Test.Embedded.X", [[ return {Name = "X"} ]])
+vfs:AddFile("Test.Embedded.Y", [[ return {Name = "Y", Friend = (...).X} ]])
+
+core:AddRebase("^Test%.Embedded%.", N.Test.Embedded)
+
+-- FullyLoad capability
 local Test = N.Test
 Test:FullyLoad()
 
 assert(N.HelloFlag, "FullyLoad failed to load Test.Hello")
 
-assert(N.Test, "Could not add directory 'Test'")
-assert(N.Test.Name == "Test", "_ was not loaded for 'Test'")
-assert(N.Test.Hello, "Hello was not loaded")
-assert(N.Test.Hello.Name == "Hello", "Hello did not load properly.")
+assert(Test, "Could not add directory 'Test'")
+assert(Test.Name == "Test", "_ was not loaded for 'Test'")
+assert(Test.Hello, "Hello was not loaded")
+assert(Test.Hello.Name == "Hello", "Hello did not load properly.")
+
+assert(Test.Embedded.X.Name == "X", "Rebased library did not load properly.")
+assert(Test.Embedded.Y.Friend.Name == "X", "Rebased library was not rebased correctly.")