+++ changelog.md
@@ -3,10 +3,12 @@
 ## 0.4.0
 - Case changed to PascalCase for public API.
 - Added VFS unit test, since it's a little fragile.
-- Vanilla Lua now has a non-LFS fallback
-- LOVE FS wrapper is now the highest priority
+- Vanilla Lua now has a non-LFS fallback.
+- LOVE FS wrapper is now the highest priority.
 - Parameters passed to loaded modules are now `(base, path)` instead of `(path, base)`
 - Directories now have a `FullyLoad` member function to load all members of the namespace recursively.
+- Added rebasing semantics with N:AddRebase, see the VFS unit test.
+- Clearer documentation on all internal methods.
 
 ## 0.3.1
 - Circular reference detection